+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ FORMS Feature
 * We added the feature equivalence in pykechain for KE-chain Forms and Workflow.
 * :+1: Added the `FormReferenceProperties` and tests (#1098)
 * :+1: Added the `ContextReferenceProperties` and tests (#1099)
+* :+1: Added 'Forms', 'Form models', 'Contexts' and 'Workflows' sidebar buttons in the `SideBarManager' and added ability to specify 'minimum_access_level' and 'alignment' per 'SideBarButton' (#1103)
 * :+1: Added the `Forms` class and tests (#1100)
 
 UNRELEASED
@@ -17,7 +18,6 @@ v3.17.1 (29NOV21)
 -----------------
 A re-release of v3.17.0 due to missing version information and changelog info in the library itself. No new code.
 
-v3.17.0 (29NOV21)
 -----------------
 * :+1: When a part is searched that on its `id` or `pk` than the `Client.part()` call will perform a API call to the detail route of the part api. This will greatly reduce the overhead on large databases and improve performance. (#1064)
 * :+1: dependent versions for development: sphinx (4.3.1), twine (3.6.0), coveralls (3.3.1), coverage (6.2)